[app][uefi] refactor relocation logic to a separate file
