+++ lib/uefi/relocation.h
@@ -1,0 +1,6 @@
+#ifndef __LIB_UEFI_RELOCATION_H_
+#define __LIB_UEFI_RELOCATION_H_
+
+int relocate_image(char *image);
+
+#endif
