+++ lib/uefi/rules.mk
@@ -8,6 +8,7 @@ MODULE_DEFINES=MSPACES=1
 
 MODULE_SRCS += \
 	$(LOCAL_DIR)/uefi.cpp \
+	$(LOCAL_DIR)/relocation.cpp \
 	$(LOCAL_DIR)/text_protocol.cpp \
 	$(LOCAL_DIR)/boot_service_provider.cpp \
 	$(LOCAL_DIR)/runtime_service_provider.cpp \
